PR5: move connect_rw from scout.trino to scout.voila
trino-rw is Voila-only: NetworkPolicy allow-lists Voila-labeled pods,
and reviewer-annotation writes from per-user Jupyter kernels aren't
part of Scout's data model. Module organization now reflects that - a
Jupyter user reaching for scout.trino.connect_rw gets a clear
AttributeError instead of a TCP timeout from the NetworkPolicy.

Voila playbooks update from scout_trino.connect_rw() to
from scout.voila import connect_rw.
